Remove some IDE warning and make pattern demo.

+++ MUtilsProject/MUtils/src/main/java/cn/mutils/app/core/codec/ByteUtil.java
@@ -9,9 +9,6 @@ public class ByteUtil {
 
     /**
      * Convert integer to bytes
-     *
-     * @param value
-     * @return
      */
     public static byte[] toBytes(int value) {
         byte[] bytes = new byte[4];
@@ -24,9 +21,6 @@ public class ByteUtil {
 
     /**
      * Convert bytes to integer
-     *
-     * @param bytes
-     * @return
      */
     public static int toInt(byte[] bytes) {
         if (bytes == null) {
@@ -37,10 +31,6 @@ public class ByteUtil {
 
     /**
      * Convert bytes to integer from offset position
-     *
-     * @param bytes
-     * @param offset
-     * @return
      */
     public static int toInt(byte[] bytes, int offset) {
         if (bytes == null) {
@@ -49,18 +39,15 @@ public class ByteUtil {
         if (offset < 0 || (offset + 4) > bytes.length) {
             return 0;
         }
-        int value = ((int) (bytes[offset] & 0xFF)) << 24;
-        value |= ((int) (bytes[offset + 1] & 0xFF)) << 16;
-        value |= ((int) (bytes[offset + 2] & 0xFF)) << 8;
-        value |= ((int) (bytes[offset + 3] & 0xFF));
+        int value = (bytes[offset] & 0xFF) << 24;
+        value |= (bytes[offset + 1] & 0xFF) << 16;
+        value |= (bytes[offset + 2] & 0xFF) << 8;
+        value |= bytes[offset + 3] & 0xFF;
         return value;
     }
 
     /**
      * Convert bytes to binary string
-     *
-     * @param bytes
-     * @return
      */
     public static String toString(byte[] bytes) {
         if (bytes == null) {
@@ -71,9 +58,6 @@ public class ByteUtil {
 
     /**
      * Convert bytes to hex string
-     *
-     * @param bytes
-     * @return
      */
     public static String toHexString(byte[] bytes) {
         if (bytes == null) {
